Organized pages and components and minor fixes
