--- client/.env.example
@@ -1,1 +1,0 @@
-GOOGLE_MAPS_API_KEY =
