+++ client/jsconfig.json
@@ -1,0 +1,9 @@
+{
+    "compilerOptions": {
+        "baseUrl": "src",
+        "paths": {
+            "components/*": ["components/*"]
+        }
+    },
+    "include": ["src"]
+}
