+++ client/src/components/Navbar/index.jsx
@@ -8,43 +8,36 @@ import Chip from '@mui/material/Chip';
 import Stack from '@mui/material/Stack';
 
 const Navbar = () => {
-  const handleHomeClick = () => {
-    // alert("Home clicked");
-  }
   const handleSearchPress = (event) => {
-    if(event.key === 'Enter'){
+    if (event.key === 'Enter') {
       alert("Enter pressed");
     }
   }
-  const handlePostClick = () => {
-   // alert("Post clicked");
-  }
-  const handleProfileClick = () => {
-    // alert("Profile clicked");
-  }
   const handleLogoutClick = () => {
-    // alert("Logout clicked");
     localStorage.removeItem('user');
   }
 
-  
+
   return (
     <div className="navbar">
       <nav className="navlinks">
         <a href="/home">
-          <FoodBankOutlinedIcon onClick={handleHomeClick} style={{ fontSize: '3rem' }} />
+          <div className="navbar-logo">
+          <FoodBankOutlinedIcon style={{ fontSize: '3rem' }} />
+          <h3>IGFood</h3>
+          </div>
         </a>
         <div className="search-container">
-          <Box component="form" sx={{'& > :not(style)': { m: 1, width: '25ch' },}} noValidate autoComplete="off">
+          <Box component="form" sx={{ '& > :not(style)': { m: 1, width: '25ch' }, }} noValidate autoComplete="off">
             <TextField input="text" id="filled-basic" label="Search" variant="filled" onKeyPress={handleSearchPress} /> {/* not implemented */}
           </Box>
         </div>
-        <Stack direction="row" spacing={3} sx={{marginLeft: "3%"}}>
+        <Stack direction="row" spacing={3} sx={{ marginLeft: "3%" }}>
           <a href="/post">
-            <Chip label="Make a Post" onClick={handlePostClick} />
+            <Chip label="Make a Post" clickable/>
           </a>
           <a href="/profile/:userID">
-            <Chip label="My Profile" onClick={handleProfileClick} />
+            <Chip label="My Profile" clickable/>
           </a>
         </Stack>
         <a href="/">
