+++ client/src/components/StarRating/index.jsx
