+++ client/src/pages/SignupPage/index.jsx
@@ -1,22 +1,23 @@
-import React, {useState} from "react";
+import React, { useState } from "react";
 import { useNavigate } from 'react-router-dom';
+import restaurant from 'assets/restaurant.jpeg';
+// MUI Components
 import FoodBankOutlinedIcon from '@mui/icons-material/FoodBankOutlined';
-import IconButton from '@mui/material/IconButton'
-import "./signupPage.css";
-import { FormControl, InputAdornment, InputLabel, OutlinedInput, TextField, Button, Link } from "@mui/material";
 import Visibility from '@mui/icons-material/Visibility'
 import VisibilityOff from "@mui/icons-material/Visibility";
-import restaurant from './restaurant.jpeg';
-import API from "../API";
+import { FormControl, InputAdornment, InputLabel, 
+    OutlinedInput, TextField, Button, Link, IconButton } from "@mui/material";
+import API from "api/API";
 import bcrypt from 'bcryptjs';
+import "./SignupPage.css";
 
-//NEED TO ADD LOGO At toP, HOW??
+
 const SignupPage = () => {
     const [email, setEmail] = useState("");
     const [name, setName] = useState("");
     const [password, setPassword] = useState("");
     const [showPassword, setShowPassword] = useState(false);
-    const handleClickShowPassword = () => setShowPassword( (show) => !show);
+    const handleClickShowPassword = () => setShowPassword((show) => !show);
     const handleMouseDownPassword = (event) => {
         event.preventDefault();
     };
@@ -34,58 +35,60 @@ const SignupPage = () => {
         navigate("/home")
     };
 
-    return(
+    return (
         <form>
             <div className="food_image">
-            <img src={restaurant} alt="Logo" />
+                <img src={restaurant} alt="Logo" />
             </div>
             <div className="signup-form">
-                <div className = "foodIcon">
+                <div className="foodIcon">
                     <FoodBankOutlinedIcon sx={{ fontSize: "10rem" }}></FoodBankOutlinedIcon>
                     <h1 id="app_name">IGFood</h1>
                 </div>
                 <h1 id="createAccount">Create your account</h1>
                 <div className="create-row">
-                    <TextField id="outlined-basic" sx={{ width: "60%", backgroundColor: "#F5F5F5" }} 
-                    label="Full name" variant="outlined"
-                    value={name} onChange={(e) => setName(e.target.value)}/>   
+                    <TextField id="outlined-basic" sx={{ width: "60%", backgroundColor: "#F5F5F5" }}
+                        label="Full name" variant="outlined"
+                        value={name} onChange={(e) => setName(e.target.value)} />
                 </div>
                 <div className="create-row">
-                    <TextField id="outlined-basic" sx={{ width: "60%", backgroundColor: "#F5F5F5" }} 
-                    label="Email Address" variant="outlined"
-                    value={email} type="email" onChange={(e) => setEmail(e.target.value)}/>
+                    <TextField id="outlined-basic" sx={{ width: "60%", backgroundColor: "#F5F5F5" }}
+                        label="Email Address" variant="outlined"
+                        value={email} type="email" onChange={(e) => setEmail(e.target.value)} />
                 </div>
                 <div className="create-row">
-                    <FormControl sx={{m: 1, width: "60%", backgroundColor: "#F5F5F5"}} variant="outlined">
+                    <FormControl sx={{ m: 1, width: "60%", backgroundColor: "#F5F5F5" }} variant="outlined">
                         <InputLabel htmlFor="outlined-adornment-password">Enter Password</InputLabel>
-                        <OutlinedInput 
+                        <OutlinedInput
                             id="outlined_adornment_password"
                             type={showPassword ? 'text' : 'password'}
                             endAdornment={
                                 <InputAdornment position="end">
                                     <IconButton aria-label="toggle password visibility"
-                                    onClick={handleClickShowPassword}
-                                    onMouseDown={handleMouseDownPassword}
-                                    edge="end"
+                                        onClick={handleClickShowPassword}
+                                        onMouseDown={handleMouseDownPassword}
+                                        edge="end"
                                     >
-                                        {showPassword ? <VisibilityOff /> : <Visibility/>}
+                                        {showPassword ? <VisibilityOff /> : <Visibility />}
                                     </IconButton>
                                 </InputAdornment>
                             }
                             label="password"
                             value={password} onChange={(e) => setPassword(e.target.value)}
-                            />
+                        />
                     </FormControl>
                 </div>
                 <Button id="create_account_bttn" variant="contained" onClick={handleCreateUser}>Create your account</Button>
-                <Link sx={{display: "block", 
-                marginTop: "5%", 
-                fontSize: "18px", 
-                fontWeight: 400,
-                color: "#007AFF"}} underline="none" href="/">Back</Link>
+                <Link sx={{
+                    display: "block",
+                    marginTop: "5%",
+                    fontSize: "18px",
+                    fontWeight: 400,
+                    color: "#007AFF"
+                }} underline="none" href="/">Back</Link>
             </div>
         </form>
-        
+
     );
 };
